Move direct-funding state into channel-state
The reasoning here is that the channel-state branch of the redux store
will serve as the source of truth for the properties of a channel.

This includes the latest commitments, as well as the direct deposits
against that channel.

The various processes/procedures have their own branch of state, which
stores the current state of that process. These processes refer to the
channel-state branch for the current properties of that channel.
