+++ packages/wallet/src/redux/__tests__/initialized.test.ts
@@ -1,7 +1,6 @@
 import { walletReducer } from '../reducer';
 
 import * as states from './../state';
-import * as fundingStates from './../funding-state/state';
 import * as actions from './../actions';
 
 const defaults = {
@@ -23,14 +22,5 @@ describe('when the player initializes a channel', () => {
     expect(updatedState.channelState.initializingChannels[ids[0]].privateKey).toEqual(
       expect.any(String),
     );
-  });
-});
-
-describe.skip('when a funding related action arrives', () => {
-  const action = actions.funding.fundingReceivedEvent('0xf00', '0x', '0x');
-  const updatedState = walletReducer(initializedState, action);
-
-  it('applies the funding state reducer', async () => {
-    expect(updatedState.fundingState).toEqual(fundingStates.FUNDING_NOT_STARTED);
   });
 });
