+++ packages/wallet/src/redux/channel-state/direct-funding/__tests__/direct-funding.test.ts
@@ -18,7 +18,6 @@ const YOUR_DEPOSIT_B = twoThree[0];
 const TOTAL_REQUIRED = twoThree.reduce(addHex);
 
 const defaultsForA: states.DirectFundingStatus = {
-  fundingType: states.DIRECT_FUNDING,
   requestedTotalFunds: TOTAL_REQUIRED,
   requestedYourContribution: YOUR_DEPOSIT_A,
   channelId,
