+++ packages/wallet/src/redux/channel-state/direct-funding/depositing/__tests__/depositing.test.ts
@@ -23,7 +23,6 @@ const YOUR_DEPOSIT_B = bigNumberify(TOTAL_REQUIRED)
   .toHexString();
 
 const defaultsForA: states.Depositing = {
-  fundingType: directFundingStates.DIRECT_FUNDING,
   requestedTotalFunds: TOTAL_REQUIRED,
   requestedYourContribution: YOUR_DEPOSIT_A,
   channelId,
