+++ packages/wallet/src/redux/channel-state/direct-funding/depositing/reducer.ts
