+++ packages/wallet/src/redux/channel-state/direct-funding/depositing/state.ts
