+++ packages/wallet/src/redux/reducer.ts
@@ -6,7 +6,6 @@ import { clearOutbox } from './outbox/reducer';
 import { accumulateSideEffects } from './outbox';
 import { initializationSuccess } from 'magmo-wallet-client/lib/wallet-events';
 import { channelStateReducer } from './channel-state/reducer';
-import { fundingStateReducer } from './funding-state/reducer';
 import { combineReducersWithSideEffects } from './../utils/reducer-utils';
 
 const initialState = states.waitForLogin();
@@ -49,7 +48,6 @@ export function initializedReducer(
 
 const combinedReducer = combineReducersWithSideEffects({
   channelState: channelStateReducer,
-  fundingState: fundingStateReducer,
 });
 
 const waitForLoginReducer = (
